requirements.txt
python-telegram-bot~=20.0
openai~=1.0
google-genai~=1.0
python-dotenv~=1.0
pytest~=8.0
pytest-asyncio~=0.23
